src/requirements.txt
jupyter_client>=5.1, <7.0
PyYAML>=4.2, <6.0
rope~=0.21.0  # subject to LGPLv3 license
setuptools>=38.0
pandas==1.3.4
psycopg2==2.9.2
requests==2.26.0
pyaml==6.0
